% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_mean}
\alias{age_mean}
\title{Passenger Age Mean}
\usage{
To use this function by itself, you need to input the Titanic DF. The DF can be stored as a variable by using the function import_data(). You can then call the use the function with the command: > age_mean(df)
}
\description{
This function returns the mean age of the passengers.
}
