% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{main_func}
\alias{main_func}
\title{Main function}
\usage{
Run the function by calling it without arguments.
}
\description{
This function runs the various functions of the package in the correct order.
}
\examples{
main_func()
}
