% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_passengers}
\alias{plot_passengers}
\title{Plot Passengers}
\usage{
To use this function by itself, you need to input the Titanic DF as an argument for the function. The DF can be stored as a variable by using the function import_data(). You can then call the use the function with the command: > survivor_age_mean(df)
}
\description{
This function plots a scatter plot of the passengers, with the binary feature "Survived" on the y axis, and integer feature "Age" on the x-axis. The gender of each passenger is represented as colors. It is not included in the main-func.
}
