% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{survivor_age_mean}
\alias{survivor_age_mean}
\title{Survivor Age Mean}
\usage{
survivor_age_mean(df = NULL)
}
\arguments{
\item{df}{The Titanic dataframe. Raw data as csv from Kaggle, converted to a R dataframe with the function read_csv.}
}
\description{
This function returns the mean age of the survivors.
}
\examples{
survivor_age_mean()
}
