% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age_mean}
\alias{age_mean}
\title{Passenger Age Mean}
\usage{
To use this function by itself, you need to input the Titanic DF. The DF can be stored as a variable by using the function import_data().
}
\arguments{
\item{df}{The Titanic dataframe. Raw data as csv from Kaggle, converted to a R dataframe with the function read_csv.}
}
\description{
This function returns the mean age of the passengers.
}
\examples{
age_mean(df)
}
