% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{import_data}
\alias{import_data}
\title{Import Data}
\usage{
To use this function by itself, store the DF in a variable.
}
\description{
This function reads the data from a csv-file and returns it as a DataFrame-object.
}
\examples{
df = import_data()
}
