% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{plot_passengers}
\alias{plot_passengers}
\title{Plot Passengers}
\usage{
plot_passengers(df = NULL)
}
\arguments{
\item{df}{The Titanic dataframe. Raw data as csv from Kaggle, converted to a R dataframe with the function read_csv.}
}
\description{
This function plots a scatter plot of the passengers, with the binary feature "Survived" on the y axis, and integer feature "Age" on the x-axis. The gender of each passenger is represented as colors. It is not included in the main-func.
}
\examples{
plot_passengers()
}
